Add Email model
Email has a many-to-one relationship with User (one user can have many
emails) under the alias `:sent_to`.

Providing the User as `:sent_to` or the user_id foreign key is required
to create a valid Email entry. Did not use the email (sent_address) as
a key because users are allowed to change their emails.